# laravel-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Agency API",
        "description": "API documentation for the Travel Agency Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/hotels": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get list of hotels",
                "description": "Display a listing of the hotels.",
                "operationId": "ce2cf593cba97023cd25d7e62ee02614",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Create a new hotel",
                "description": "Store a newly created hotel in storage.",
                "operationId": "5a9226f9e5c7c492a046003a8a62b61d",
                "requestBody": {
                    "description": "Hotel data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Hotel created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hotels/{id}": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get a hotel by ID",
                "description": "Display the specified hotel.",
                "operationId": "8554f2a2fde256aef3c038ddd5dbf6b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Update a hotel",
                "description": "Update the specified hotel in storage.",
                "operationId": "885cd3264c35f072a73d81ce21819f4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Hotel data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Hotel updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Delete a hotel",
                "description": "Remove the specified hotel from storage.",
                "operationId": "9e11e1d483ce37115bcfaefc57cab373",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Hotel deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}